:recycle: refactor: extract shared @howardism/article-contract module
Centralize the article contract — the tag/topic unions, SourceRef, and the
write-side ArticleContract frontmatter shape — into one workspace package the
wiki importer consumes for types and the blog consumes for zod validation,
replacing two hand-synced mirror copies guarded only by comments.

Blog read-time-only fields (archived, dropCap, imageAlt) stay local via
ArticleContractSchema.extend. Part of the codebase-deepening series.
