+++ apps/blog/package.json
@@ -13,6 +13,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "@howardism/article-contract": "workspace:*",
     "@howardism/ui": "workspace:*",
     "@mdx-js/loader": "^3.0.1",
     "@next/mdx": "^16.2.2",
