+++ apps/cli/package.json
@@ -10,6 +10,7 @@
     "import:wiki": "bun src/import-wiki/index.ts"
   },
   "dependencies": {
+    "@howardism/article-contract": "workspace:*",
     "gray-matter": "^4.0.3",
     "yaml": "^2.9.0"
   },
