+++ packages/article-contract/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@howardism/article-contract",
+  "version": "0.0.1",
+  "private": true,
+  "license": "MIT",
+  "exports": {
+    ".": "./src/index.ts",
+    "./schema": "./src/schema.ts"
+  },
+  "scripts": {
+    "type-check": "tsc --noEmit",
+    "clean": "rm -rf node_modules"
+  },
+  "dependencies": {
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "@howardism/tsconfig": "workspace:*",
+    "typescript": "^6.0.2"
+  }
+}
