+++ packages/article-contract/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@howardism/tsconfig/base",
+  "compilerOptions": {
+    "paths": {
+      "@/*": ["./src/*"]
+    }
+  }
+}
